refactor : 3주차 시작 전 리팩토링

+++ src/main/java/kr/hhplus/be/server/interfaces/api/point/request/PointChargeRequest.java
@@ -1,5 +1,7 @@
 package kr.hhplus.be.server.interfaces.api.point.request;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "사용자 포인트 충전 요청 DTO")
@@ -14,6 +16,6 @@ public record PointChargeRequest(
 		description = "충전할 포인트, 1과 1,000,000 사이의 정수",
 		example = "10000"
 	)
-	Long chargeAmount
+	BigDecimal chargeAmount
 ) {
 }
